hw: add local core config parameter
Add local core ID to specify the core targeted by remote.
This is used only on S32xx platforms where Linux may be
running SMP and the pending interrupt needs to be cleared
in the targeted core instead of the executing core.

Also passing remote core type to be able to select a different
core index than the default.

Module parameters are now printed to a string using safer
implementation snprintf.

Issue: IPCF-1017
Signed-off-by: Manuel Arguelles <manuel.arguelles@nxp.com>